Renaming the Class1.cs to FizzBuzz.cs
I goofed and just made a Class Library instead of a console app.
Shouldn't be that big of a deal since we essentially are just running
the tests anyway. By the way, I will be using JustCode to run the tests.
Also, just so it is documented we are going to use Nunit as are unit
testing framework.